* commit '3fe9449683ded379549138e4d0e563970e70000a':
  「救援， 协助」页面修复，完善。
  救援 / 协办添加（未完成）

+++ ios/XMDD/Model/GetRescueHistoryOp.h
@@ -9,7 +9,7 @@
 #import "BaseOp.h"
 
 @interface GetRescueHistoryOp : BaseOp
-@property (nonatomic, assign) long long  applytime;
+@property (nonatomic, assign) NSUInteger applytime;
 @property (nonatomic, assign) NSInteger  type;
 @property (nonatomic, strong) NSArray *rsp_applysecueArray;
 @end
